Problem: hex_directions is untested and turns out, buggy
Solution:

Write simple unit tests, fix the bug.
Next up for me is to add a propery-based test here both as practice and to make
sure that invariants of adj functions stand across the board.

I also need to cover all the public functions with test in this module, but
prop testing first!